+++ src/libs/decoders/mp3_seek_table.cpp
@@ -142,7 +142,7 @@ Uint64 calculate_stream_hash(struct SDL_RWops* const context) {
     // Seek to the middle of the file while taking into account version small files.
     const Sint64 tail_size = (stream_size > 32768) ? 32768 : stream_size;
     const Sint64 mid_pos = static_cast<Sint64>(stream_size/2.0) - tail_size;
-    SDL_RWseek(context, mid_pos >= 0 ? mid_pos : 0, RW_SEEK_SET);
+    SDL_RWseek(context, mid_pos >= 0 ? (int)mid_pos : 0, RW_SEEK_SET);
 
     // Prepare our read buffer and counter:
     vector<char> buffer(1024, 0);
@@ -163,7 +163,7 @@ Uint64 calculate_stream_hash(struct SDL_RWops* const context) {
 
     while (total_bytes_read < static_cast<size_t>(tail_size)) {
         // Read a chunk of data.
-        const size_t bytes_read = SDL_RWread(context, buffer.data(), 1, buffer.size());
+        const size_t bytes_read = SDL_RWread(context, buffer.data(), 1, (int)buffer.size());
 
         if (bytes_read != 0) {
             // Update our hash if we read data.
@@ -175,7 +175,7 @@ Uint64 calculate_stream_hash(struct SDL_RWops* const context) {
     }
 
     // restore the stream position
-    SDL_RWseek(context, original_pos, RW_SEEK_SET);
+    SDL_RWseek(context, (int)original_pos, RW_SEEK_SET);
 
     const Uint64 hash = XXH64_digest(state);
     XXH64_freeState(state);
