+++ src/libs/decoders/wav.c
@@ -55,7 +55,7 @@ static size_t wav_read(void* pUserData, void* pBufferOut, size_t bytesToRead)
     size_t bytes_read = 0;
 
     while (bytes_read < bytesToRead) {
-        const size_t rc = SDL_RWread(rwops, ptr, 1, bytesToRead - bytes_read);
+        const size_t rc = SDL_RWread(rwops, ptr, 1, (int)(bytesToRead - bytes_read));
         if (rc == 0) {
             sample->flags |= SOUND_SAMPLEFLAG_EOF;
             break;
